Display dependency reasons in dependency report task
This commit adds support for showing the dependency (constraint) reasons in the dependency report,
in addition to the dependency insight report which was previously implemented. The report will only
show something if at least one custom reason is provided. In other words, it doesn't change the fact
that it doesn't show "conflict resolution", for example. Only the last registered description is going
to be displayed.

Signed-off-by: Cedric Champeau <cedric@gradle.com>